refactor(db): consolidate seed scripts with incremental seeding
- Update seed-safe.ts to add only new foods when FORCE_SEED=false
- Add Redis cache invalidation to seed-safe.ts
- Remove seed.ts (functionality merged into seed-safe.ts)
- Update package.json: 'seed' for incremental, 'seed:force' for full reset

New behavior:
- pnpm seed: adds only new foods (keeps existing IDs)
- pnpm seed:force: drops and reseeds everything
